requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
docstring_parser==0.16
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
grpcio==1.71.0
huggingface-hub==0.30.2
idna==3.10
importlib_resources==6.5.2
isort==6.0.1
Jinja2==3.1.6
jsonargparse==4.38.0
kiwisolver==1.4.8
lightning==2.5.1
lightning-utilities==0.14.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pathspec==0.12.1
pillow==11.2.1
pip==25.0.1
platformdirs==4.3.7
propcache==0.3.1
protobuf==6.30.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
setuptools==78.1.0
shtab==1.7.2
six==1.17.0
standard-imghdr==3.13.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchmetrics==1.7.1
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.3
typeguard==4.4.2
typeshed_client==2.7.0
typing_extensions==4.13.2
tyro==0.9.19
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
